Apply slightly modified version of patches for HARMONY-796 ([classlib][nio]java.nio.ByteBuffer.wrap(byte[], int off, int len) does not throw IndexOutOfBoundsException if off/len ==Integer.MAX_VALUE)
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@419909 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/nio/src/main/java/java/nio/ByteBuffer.java
@@ -104,7 +104,7 @@ public abstract class ByteBuffer extends Buffer implements Comparable<ByteBuffer
                 if (array == null) {
                         throw new NullPointerException();
                 }
-                if ((start< 0 ) || (len < 0) || start+ len > array.length) {
+                if ((start < 0 ) || (len < 0) || ((long)start + (long)len > array.length)) {
                         throw new IndexOutOfBoundsException();
                 }
 
@@ -427,7 +427,7 @@ public abstract class ByteBuffer extends Buffer implements Comparable<ByteBuffer
 	 *                <code>remaining()</code>
 	 */
 	public ByteBuffer get(byte[] dest, int off, int len) {
-        if ((off < 0 ) || (len < 0) || off + len > dest.length) {
+        if ((off < 0 ) || (len < 0) || ((long)off + (long)len > dest.length)) {
             throw new IndexOutOfBoundsException();
         }
         
@@ -784,7 +784,7 @@ public abstract class ByteBuffer extends Buffer implements Comparable<ByteBuffer
 	 *                If no changes may be made to the contents of this buffer
 	 */
 	public ByteBuffer put(byte[] src, int off, int len) {
-        if ((off < 0 ) || (len < 0) || off + len > src.length) {
+        if ((off < 0 ) || (len < 0) || ((long)off + (long)len > src.length)) {
             throw new IndexOutOfBoundsException();
         }
